refact: move properties file to properties dir
