+++ src/main/kotlin/com/dobby/backend/DobbyBackendApplication.kt
@@ -1,8 +1,6 @@
 package com.dobby.backend
 
-import com.dobby.backend.infrastructure.config.TokenProperties
 import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.boot.runApplication
 
 @SpringBootApplication
