+++ src/main/kotlin/com/dobby/backend/config/ApplicationConfig.kt
@@ -1,7 +1,7 @@
 package com.dobby.backend.config
 
-import com.dobby.backend.infrastructure.config.S3Properties
-import com.dobby.backend.infrastructure.config.TokenProperties
+import com.dobby.backend.infrastructure.config.properties.S3Properties
+import com.dobby.backend.infrastructure.config.properties.TokenProperties
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Configuration
