+++ src/main/kotlin/com/dobby/backend/infrastructure/config/properties/S3Properties.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.infrastructure.config
+package com.dobby.backend.infrastructure.config.properties
 
 import org.springframework.boot.context.properties.ConfigurationProperties
