+++ src/main/kotlin/com/dobby/backend/infrastructure/config/properties/TokenProperties.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.infrastructure.config
+package com.dobby.backend.infrastructure.config.properties
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.stereotype.Component
