+++ src/main/kotlin/com/dobby/backend/infrastructure/token/JwtTokenProvider.kt
@@ -1,7 +1,7 @@
 package com.dobby.backend.infrastructure.token
 
 import com.dobby.backend.domain.exception.*
-import com.dobby.backend.infrastructure.config.TokenProperties
+import com.dobby.backend.infrastructure.config.properties.TokenProperties
 import io.jsonwebtoken.ExpiredJwtException
 import io.jsonwebtoken.JwtException
 import io.jsonwebtoken.Jwts
